+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Recommendation: for ease of reading, use the following order:
 - Fixed
 -->
 
+## [Unreleased]
+### Fixed
+- Flow APIs: correctly returns response for dataset with cleaned blocks 
+
 ## [0.229.0] - 2025-03-24
 ### Added
 - Experimental support for `ChangelogStream` and `UpsertStream` ODF merge strategies.